Add basictest.py to project (for testing)
The test is currently only for old versions (without USART extension)!
It sends 5 'a' characters over serial connection and
expect that at least one character arrive on the other side.
The timer are needed to verify that it works, so dont delete the
timer.sleep() until u know what u'r doing.