Adding "offset counter" to layer1
At the moment the logic behind the module is not really changed.
Its the way that the frequency is increased and a counter within the
layer1 interrupt is the counterpart of that increase.
--> NO LOGIC IS CHANGED!
It will follow the calculation of a random offset for the
layer1 receiver, so that we can find the rhythm of the sender.